refactor: Controller의 run()에서 모든 예외를 잡도록 변경. 예외 발생시 프로그램이 종료되도록 함.
