+++ src/test/java/lotto/domain/LottoNumberTest.java
@@ -26,12 +26,4 @@ public class LottoNumberTest {
             .isThrownBy(() -> new LottoNumber(46));
     }
 
-    @DisplayName("당첨 로또는 보너스볼과 숫자가 중복되어선 안된다.")
-    @Test
-    void duplicate(){
-        LottoNumber bonusBall = new LottoNumber(3);
-        LottoTicket ticket = new LottoTicket(Arrays.asList(1,2,3,4,5,6));
-        assertThatIllegalArgumentException()
-            .isThrownBy(()->bonusBall.validateDuplicate(ticket));
-    }
 }
